Home

PkgOptions.jl

PkgOptions.load

PkgOptions.rm

PkgOptions.set

PkgOptions.load — Function.

PkgOptions.load(pkg::Symbol)

source

PkgOptions.rm — Function.

PkgOptions.rm(pkg::Symbol)

Remove package options (use the defaults) of package pkg.

source

PkgOptions.set — Function.

PkgOptions.set(pkg::Symbol, options::Pair...)
PkgOptions.set(pkg::Symbol; options...)

Set and/or add options to package options for package pkg.

PkgOptions.set(pkg::Symbol, options::AbstractDict)

Set package options of package pkg to options, removing the ones not specified by options.

source